requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.14.2
chardet==3.0.4
Click==7.0
cryptography==2.3.1
enum34==1.1.6
Flask==1.0.2
future==0.16.0
idna==2.7
influxdb==5.2.0
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
pycparser==2.19
pymongo==3.7.1
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-twitter==3.4.2
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.2
Werkzeug==0.15.5
